* Whoops, these are floats, not uint's, fixes assertion error in cellhandler.h.
git-svn-id: svn://svn.aspiredev.org/AspireCore/branches/Hearthstone@1060 a6a5f009-272a-4b40-a74d-5f9816a51f88

+++ src/hearthstone-world/MapMgr.cpp
@@ -280,8 +280,8 @@ void MapMgr::PushObject(ObjectPointer obj)
 	ASSERT(obj->GetPositionY() < _maxY && obj->GetPositionY() > _minY);
 	ASSERT(_cells);
 
-	uint32 mx = obj->GetPositionX();
-	uint32 my = obj->GetPositionY();
+	float mx = obj->GetPositionX();
+	float my = obj->GetPositionY();
 	uint32 cx = GetPosX(mx);
 	uint32 cy = GetPosY(my);
